Split settings per card
